# docs/swagger.yaml
basePath: /
definitions:
...
  credentials.Credentials:
    properties:
      login:
        example: someLogin
        type: string
      password:
        example: somePassword
        type: string
      site_id:
        $ref: '#/definitions/sites.Site'
    type: object
  sites.Site:
    properties:
      host:
        example: site.com
        type: string
      name:
        example: site
        type: string
    type: object
host: localhost:80
info:
  contact: {}
  description: How to work CRUD server in reverseProxy
  title: CRUD server in reverseProxy
  version: 1.0.0
paths:
...
  /credentials:
    post:
      consumes:
      - application/json
      description: Create credentials
      parameters:
      - description: credentials info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/credentials.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: string
      summary: Create new credentials
      tags:
      - Credentials
  /credentials/{id}:
    delete:
      consumes:
      - application/json
      description: delete credentials
      parameters:
      - description: credentials ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credentials.Credentials'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete credentials based on given id
      tags:
      - Credentials
    get:
      consumes:
      - application/json
      description: get credentials
      parameters:
      - description: credentials ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credentials.Credentials'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get credentials based on given id
      tags:
      - Credentials
    put:
      consumes:
      - application/json
      description: update credentials
      parameters:
      - description: credential ID
        in: path
        name: id
        required: true
        type: integer
      - description: credentials info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/credentials.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credentials.Credentials'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Update credentials based on given id
      tags:
      - Credentials
...
schemes:
- http
swagger: "2.0"